Rename schema classes
