+++ plugins/woocommerce/src/Blocks/BlockTypes/Checkout.php
@@ -1,11 +1,12 @@
 <?php
+declare( strict_types = 1);
 namespace Automattic\WooCommerce\Blocks\BlockTypes;
 
 use Automattic\WooCommerce\StoreApi\Utilities\LocalPickupUtils;
 use Automattic\WooCommerce\Blocks\Utils\CartCheckoutUtils;
 use Automattic\WooCommerce\Blocks\Package;
 use Automattic\WooCommerce\Blocks\Domain\Services\CheckoutFields;
-use Automattic\WooCommerce\Blocks\Utils\CheckoutFieldsSchema;
+use Automattic\WooCommerce\Blocks\Domain\Services\CheckoutFieldsSchema\Validation;
 use Automattic\WooCommerce\Admin\Features\Features;
 
 /**
@@ -166,7 +167,7 @@ class Checkout extends AbstractBlock {
 		if ( Features::is_enabled( 'experimental-blocks' ) ) {
 			$checkout_fields = Package::container()->get( CheckoutFields::class );
 			// Load schema parser asynchronously if we need it.
-			if ( CheckoutFieldsSchema::has_field_with_schema( $checkout_fields->get_additional_fields() ) ) {
+			if ( Validation::has_field_with_schema( $checkout_fields->get_additional_fields() ) ) {
 				$dependencies[] = 'wc-schema-parser';
 			}
 		}
