+++ plugins/woocommerce/src/Blocks/Domain/Services/CheckoutFieldsSchema/Validation.php
@@ -1,20 +1,20 @@
 <?php
 declare( strict_types = 1);
 
-namespace Automattic\WooCommerce\Blocks\Utils;
+namespace Automattic\WooCommerce\Blocks\Domain\Services\CheckoutFieldsSchema;
 
-use Automattic\WooCommerce\Blocks\Domain\Services\Schema\DocumentObject;
-use Opis\JsonSchema\Helper;
-use Opis\JsonSchema\Validator;
-use Opis\JsonSchema\Errors\{
-	ErrorFormatter,
-	ValidationError,
+use Automattic\WooCommerce\Blocks\Domain\Services\CheckoutFieldsSchema\DocumentObject;
+use Opis\JsonSchema\{
+	Helper,
+	Validator
 };
+use Opis\JsonSchema\Errors\ErrorFormatter;
+use WP_Error;
 
 /**
- * Service class managing checkout field schema.
+ * Service class validating checkout field schema.
  */
-class CheckoutFieldsSchema {
+class Validation {
 	/**
 	 * Meta schema.
 	 *
@@ -55,7 +55,7 @@ class CheckoutFieldsSchema {
 	 *
 	 * @param DocumentObject $document_object The document object to validate.
 	 * @param array          $rules The rules to validate against.
-	 * @return bool|\WP_Error
+	 * @return bool|WP_Error
 	 */
 	public static function validate_document_object( DocumentObject $document_object, $rules ) {
 		$validator = new Validator();
@@ -79,7 +79,7 @@ class CheckoutFieldsSchema {
 		$formatted = $formatter->formatFlat( $result->error() );
 		$errors    = implode( ', ', count( $formatted ) > 2 ? array_slice( $formatted, 2 ) : $formatted );
 
-		return new \WP_Error( 'woocommerce_rest_checkout_invalid_field', $errors );
+		return new WP_Error( 'woocommerce_rest_checkout_invalid_field', $errors );
 	}
 
 	/**
@@ -112,12 +112,12 @@ class CheckoutFieldsSchema {
 	 * Validate meta schema for field rules.
 	 *
 	 * @param array $rules The field rules.
-	 * @return bool|\WP_Error True if the field options are valid, a WP_Error otherwise.
+	 * @return bool|WP_Error True if the field options are valid, a WP_Error otherwise.
 	 */
 	public static function is_valid_schema( $rules ) {
 		if ( empty( self::$meta_schema_json ) ) {
 			// phpcs:ignore WordPress.WP.AlternativeFunctions.file_get_contents_file_get_contents
-			self::$meta_schema_json = file_get_contents( __DIR__ . '/Schema/json-schema-draft-07.json' );
+			self::$meta_schema_json = file_get_contents( __DIR__ . '/json-schema-draft-07.json' );
 		}
 		$validator    = new Validator();
 		$test_schemas = [ 'required', 'hidden', 'validation' ];
@@ -127,7 +127,7 @@ class CheckoutFieldsSchema {
 				continue;
 			}
 			if ( ! is_array( $rules[ $rule ] ) ) {
-				return new \WP_Error( 'woocommerce_rest_checkout_invalid_field_schema', sprintf( 'The %s rules must be an array.', esc_html( $rule ) ) );
+				return new WP_Error( 'woocommerce_rest_checkout_invalid_field_schema', sprintf( 'The %s rules must be an array.', esc_html( $rule ) ) );
 			}
 			$result = $validator->validate(
 				Helper::toJSON(
@@ -143,7 +143,7 @@ class CheckoutFieldsSchema {
 				self::$meta_schema_json
 			);
 			if ( $result->hasError() ) {
-				return new \WP_Error( 'woocommerce_rest_checkout_invalid_field_schema', esc_html( (string) $result->error() ) );
+				return new WP_Error( 'woocommerce_rest_checkout_invalid_field_schema', esc_html( (string) $result->error() ) );
 			}
 		}
 		return true;
